// src/app/shared/models/ingresos/totalporcentajes.model.ts
export class TotalPorcentajes {

    totalE: number = 0;
    anioAntE: number = 0;
    presupuestoE: number = 0;
    proyeccionE: number = 0;
    aniATotalE: number = 0;
    presTotalE: number = 0;
    ProyTotalE: number = 0;
  
    totalFB: number = 0;
    anioAntFB: number = 0;
    presupuestoFB: number = 0;
    proyeccionFB: number = 0;
    aniATotalFB: number = 0;
    presTotalFB: number = 0;
    ProyTotalFB: number = 0;
  
    totalM: number = 0;
    anioAntM: number = 0;
    presupuestoM: number = 0;
    proyeccionM: number = 0;
    aniATotalM: number = 0;
    presTotalM: number = 0;
    ProyTotalM: number = 0;
  
    totalA: number = 0;
    anioAntA: number = 0;
    presupuestoA: number = 0;
    proyeccionA: number = 0;
    aniATotalA: number = 0;
    presTotalA: number = 0;
    ProyTotalA: number = 0;
  
    totalMY: number = 0;
    anioAntMY: number = 0;
    presupuestoMY: number = 0;
    proyeccionMY: number = 0;
    aniATotalMY: number = 0;
    presTotalMY: number = 0;
    ProyTotalMY: number = 0;
  
    totalJN: number = 0;
    anioAntJN: number = 0;
    presupuestoJN: number = 0;
    proyeccionJN: number = 0;
    aniATotalJN: number = 0;
    presTotalJN: number = 0;
    ProyTotalJN: number = 0;
  
    totalJL: number = 0;
    anioAntJL: number = 0;
    presupuestoJL: number = 0;
    proyeccionJL: number = 0;
    aniATotalJL: number = 0;
    presTotalJL: number = 0;
    ProyTotalJL: number = 0;

    totalAG: number = 0;
    anioAntAG: number = 0;
    presupuestoAG: number = 0;
    proyeccionAG: number = 0;
    aniATotalAG: number = 0;
    presTotalAG: number = 0;
    ProyTotalAG: number = 0;

    totalS: number = 0;
    anioAntS: number = 0;
    presupuestoS: number = 0;
    proyeccionS: number = 0;
    aniATotalS: number = 0;
    presTotalS: number = 0;
    ProyTotalS: number = 0;

    totalOC: number = 0;
    anioAntOC: number = 0;
    presupuestoOC: number = 0;
    proyeccionOC: number = 0;
    aniATotalOC: number = 0;
    presTotalOC: number = 0;
    ProyTotalOC: number = 0;

    totalNV: number = 0;
    anioAntNV: number = 0;
    presupuestoNV: number = 0;
    proyeccionNV: number = 0;
    aniATotalNV: number = 0;
    presTotalNV: number = 0;
    ProyTotalNV: number = 0;

    totalDC: number = 0;
    anioAntDC: number = 0;
    presupuestoDC: number = 0;
    proyeccionDC: number = 0;
    aniATotalDC: number = 0;
    presTotalDC: number = 0;
    ProyTotalDC: number = 0;
}